src/data/usecase/account/db_add_account.ts
import { AccountModel, AddAccount, AddAccountModel, AddAccountRepository, Hasher } from './db_add_account_protocols'

export class DbAddAccount implements AddAccount {
  constructor (private readonly hasher: Hasher, private readonly addAccountRepository: AddAccountRepository) { }

  async add (accountData: AddAccountModel): Promise<AccountModel> {
    const hashedPasswd = await this.hasher.hash(accountData.password)
    const account = await this.addAccountRepository.add(Object.assign({}, accountData, { password: hashedPasswd }))
    return account
  }
}
